requirements.txt
z3-solver>=4.12.0
pydantic>=2.0
streamlit>=1.30.0
plotly>=5.18.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytest>=8.0.0
mypy>=1.8.0
ruff>=0.2.0
